New test classes and //TODOs added
