+++ src/main/account/AccountList.java
@@ -3,7 +3,5 @@ package main.account;
 import java.util.ArrayList;
 
 public class AccountList {
-
     public static ArrayList<Account> accountArrayList = new ArrayList<>();
-
 }
