+++ src/test/unit_account/CreateAccountTest.java
@@ -1,0 +1,11 @@
+package test.unit_account;
+
+import org.junit.Test;
+
+public class CreateAccountTest {
+
+    @Test
+    public void testCreateAccount() {
+
+    }
+}
