+++ src/test/unit_account/LoginTest.java
@@ -1,6 +1,5 @@
 package test.unit_account;
 
-import main.GetInputs;
 import main.Login;
 import main.account.Account;
 import main.account.AccountList;
@@ -12,7 +11,6 @@ import org.junit.Test;
 
 public class LoginTest {
     Login login = new Login();
-    GetInputs getInput = new GetInputs();
     UserModel user = new UserModel(10, 0, Role.USER, Department.DEVELOPER, "user1", "qwerty");
     Account account = user;
 
@@ -24,17 +22,7 @@ public class LoginTest {
     }
 
     @Test
-    public void testUsernameInput() {
-        Assert.assertEquals("Incorrect username returned", "user", getInput.usernameInput("user"));
-    }
-
-    @Test
-    public void testPasswordInput() {
-        Assert.assertEquals("Incorrect password returned", "password", getInput.usernameInput("password"));
-    }
-
-    @Test
     public void testLaunchMainMenu() {
-
+        //TODO
     }
 }
